+++ valhalla/proto_conversions.h
@@ -33,7 +33,8 @@ inline TripLeg_Surface GetTripLegSurface(const baldr::Surface surface) {
 // TODO - why doesn't these use an enum input?
 constexpr VehicleType kTripLegVehicleType[] = {
     VehicleType::kCar,   VehicleType::kMotorcycle,   VehicleType::kAutoBus,
-    VehicleType::kTruck, VehicleType::kMotorScooter, VehicleType::kGolfCart,
+    VehicleType::kTruck, VehicleType::kMotorScooter, VehicleType::kLowSpeedVehicle,
+    VehicleType::kGolfCart,
 };
 inline VehicleType GetTripLegVehicleType(const uint8_t type) {
   return (type <= static_cast<uint8_t>(sif::VehicleType::kGolfCart)) ? kTripLegVehicleType[type]