+++ valhalla/sif/lowspeedvehiclecost.h
@@ -1,5 +1,5 @@
-#ifndef VALHALLA_GOLFCARTCOST_H
-#define VALHALLA_GOLFCARTCOST_H
+#ifndef VALHALLA_LOWSPEEDVEHICLECOST_H
+#define VALHALLA_LOWSPEEDVEHICLECOST_H
 
 #include <cstdint>
 
@@ -12,24 +12,24 @@ namespace valhalla {
 namespace sif {
 
 /**
- * Parses the golf_cart cost options from json and stores values in pbf.
+ * Parses the low_speed_vehicle cost options from json and stores values in pbf.
  * @param doc The json request represented as a DOM tree.
  * @param costing_options_key A string representing the location in the DOM tree where the costing
  *                            options are stored.
  * @param pbf_costing A mutable protocol buffer where the parsed json values will be stored.
  */
-void ParseGolfCartCostOptions(const rapidjson::Document& doc,
-                              const std::string& costing_options_key,
-                              Costing* pbf_costing);
+void ParseLowSpeedVehicleCostOptions(const rapidjson::Document& doc,
+                                     const std::string& costing_options_key,
+                                     Costing* pbf_costing);
 
 /**
- * Create golf cart cost method.
+ * Create low-speed vehicle costing method.
  * @param  options pbf with request options.
  */
-cost_ptr_t CreateGolfCartCost(const Costing& costing);
+cost_ptr_t CreateLowSpeedVehicleCost(const Costing& costing);
 
 } // namespace sif
 } // namespace valhalla
 
 
-#endif // VALHALLA_GOLFCARTCOST_H
+#endif // VALHALLA_LOWSPEEDVEHICLECOST_H